"refactor(profiles-and-preferences-management:domain): align profile queries for portfolio and user info Context: queries for portfolio items and profile info had inconsistencies in naming and usage. Decision: standardized query classes for better readability and maintainability. Impact: no behavior change, improves consistency across query handlers."

+++ Profiles/Domain/Model/Queries/GetProfileInfoByUserIdQuery.cs
@@ -1,3 +1,3 @@
 ﻿namespace Hampcoders.Electrolink.API.Profiles.Domain.Model.Queries;
 
-public record GetProfileInfoByUserIdQuery(long UserId);
+public record GetProfileInfoByUserIdQuery(int UserId);
